[GB.NET] * BUG: Make the component compile on 32 bits OS.
git-svn-id: svn://localhost/gambas/trunk@3129 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.net/src/CDnsClient.h
@@ -52,7 +52,7 @@ typedef
 	}
 	CDNSCLIENT;
 
-void dns_callback(long lParam);
+void dns_callback(intptr_t lParam);
 void* dns_get_name(void* v_obj);
 void* dns_get_ip(void* v_obj);
 void dns_event(CDNSCLIENT *mythis);
